feat: New empty project in 'client' folder
A new React project has been created inside of a 'client' folder, which
will host all of the front-end code for this project. The current app is
no more than an skeleton with no functionality.

The default project created with the `create-react-app` command has been
modified by removing content we're not going to use and creating new
folders to have the source code better organized.