feat: Comment count included in post data
The REST services to post and delete a comment have been modified to
increment and decrement a comment counter in the parent post data.

The json file containing the initial data set for the post database has
also been modified by including the new 'commentCount' property in each
of the available posts.