refactor: API help rellocated to new file
If the user access to our server's root page, they're now redirected to the
newly created 'help.html' page located in the 'public' folder.

The 'server.js' file is now shorter and easier to read, as it contains no
more than API functionality. This change also makes the help page contents
easier to maintain, as it isn't mixed up with the server's behaviour code.