feat: Redux store and initial actions
New files have been added by including an initial reducer for posts, as
well as a couple of actions to create and update a post score.

The provider component is now wrapping the whole application, so any of
the child components can now directly access Redux store.

The store is still empty and actions aren't being dispatched, but the
'App' component props contain the post list in the Redux store and the
couple of action creators defined in our current 'post.js' files.