feat: Initialize post store with API contents
Added a new fetch post actions that does retrieve the post list from the
API. The reducer updates the state based on the values returned by the
server.

Also, the 'App' component does no longer need to keep a post list on its
state: it will display the post list as it's kept in the Redux store.