feat: Display posts from category
Two new actions have been added to the app. The first one populates a new
property in the Redux store with the list of categories fetched from the
server. The second one requests for the list of posts that belongs to a
certain category.

This functionality shall be implemented in a new route, since clicking on
a category must redirect to a new page with all of the posts belonging
to that category. Right now, though, everything is still been done in the 'App'
component. Before making further progress in the app's functionality, it might be
wise to stop and apply some serious refactor to what we have now.