feat: New reducer for comments
The API helper does now contain a new method to get all of the comments
for one post. The retrieval of this data is being performed on a new
'comment.js' file in both reducers and actions folders.

This may be innecessary, though. We may include comment data as nested
information in the post list, as it's unlikely we're ever accessing
comments as an independant entity.