design: Styles for displaying post score
I've added some more new items to the css file, as well as new image
resources.

The most important change in this commit is a new score box that will be
displayed next to each post name. A little friend will show up above the
score, and it will display a different reaction depending on the current
post score. If it's a positive score, it will smile. If the score is under
zero, though, it will be angry. This adds a nice touch to the overall look
of the application, and it also makes it easy to spot reliable and good
rated posts from others that are clearly not that popular.

Also, if our little friend was not clear enough, the tag containing the
current score will be of a different color depending on the number.

This commit does also include some other minor changes in the css.