design: Enhancements in layout template and css
I've made some changes in the css file and also in the mark-up to improve
its readability and make it more maintainable. Most changes apply only to
the page header, but the rest of the css rules must be reviewed and remove
those which can be covered in more generic rules.