fix: Update post score action not to modify state
While trying out the new component I found out there was a bug in the
post reducer. The previous code for the update score action was mutating
the original post object, and so the component would not re-render after
the reducer modified the state.

Although the updatedState array was actually a new one, its items were
references to the same objects inside of the original state. Doing
'updatedState[index].voteScore' did not only update the post data in the
new array, but also in the previous state. The fact that we were
mutating that state was the reason why the component would not
re-render after the action had been dispatched.

The current commit fixes that issue.