feat + refactor: Routes for main, category and details
Some of the content in the 'App' component has been delegated into more
specific components, although there's yet work to be done.

While refactoring this code, I've added React Router components into the
mix. There's now three different routes available in the app:

 - The main page, which will display all of the available posts.

 - The category view, which will only contain the list of the post tagged
   under the specified category.

 - The post details view, where the user will be able to read all of the
   post information along with its comments.

While working on this commit, many issues arised from using React Router
along with Redux. Wrapping the 'connect' method with 'withRouter' made it
possible to navigate from one page to another, but I still had issues with
routes with parameters.

If I navigated from a category page to another, the url would be updated
in the browser but the render method wouldn't be called.

Adding a key to the rendered component appeared to somehow solve the issue, but it's very
likely we'll come accross this problem again in the future. Finding a real
solution instead of forcing the component to re-render with this patch
would be good, so we may take note of this for upcoming updates.