fix: Don't display wrong comments on details
The post details page would sometimes show the comments for a different
post when the page was visited for a second time. The cause is that the
comment box component would fetch the comments before the post was even
loaded, so the server was called with the wrong id.

This needs to be fixed in a better way by keeping a loading flag in the
redux state, so we prevent more issues like this in the future. In the
meanwhile, the comment box is not rendered unless the post id received by
props in the post details component matches the id of the post available
in the Redux store.