refactor: Generic async action creator
Since there was a lot of duplicate code among the action files, I created
a new utils file with a generic method that can be used to create an
asyncronous action.

This method will handle the 'status' and 'response' parameters, as well as
including any input property in the returned action object.

The 'createAsyncAction' function requires at least two parameters: the action type and the promise object
that would be used to dispatch the async action. The third argument is an
object with data that will be included into the resulting action object.