refactor: Changes in categories store structure
The structure for the categories store has been modified by following an
example in the Redux docs. The store contains now two properties aside
from the array with the fetched categories: a flag to signal if data is
currently being fetched and an expiration date field.

The first of the properties can be used in the application to show a
loader while data is still being read from the server. The second one will
be used to ensure the API is not called everytime a component is mounted,
and former values in the state are reused unless a certain amount of time
has passed since the last fetch call.