fix: Send date for new post in timestamp format
The server was receiving a date in a string format rather than a
timestamp, which was wrong. That caused the sort by date method not work
when trying to compare that string with an actual number representing the
creation date. With this fixed, the sort by date function is working
again.