fix: Can't create a post with empty category
If the user did not change the initial value for the category select, the
change event would not trigger. That would cause the new post to be
created with no category set, even though in the UI it would seem as it
was being added to the first category that appears selected by default.

In order to avoid that problem, I've added a disabled placeholder option
which will appear selected by default if the category prop is not set. If
the user tries to create the post without selecting one of the other items
in the select box, the form validation will fail.