feat: Edit existing post
Now it is possible to modify the title and body of an existing post.
With this, all of the required functionality for this project is
complete.

Now it's time to get serious into refactoring the code, as it's sort of messy at the
moment. Then again, first time getting into Redux, so learning through
trial and error could hardly end up in a beautiful code without making
some explicit review after having gone through all of the process and
understanding how all the pieces fit together.

There's a lot of work yet to do, but hopefully I'd be able to make this
more decent and stable before project submission.