refactor: Edit and delete buttons not shown in list
The edit and delete post buttons are now only available in the post
details page. The PostListItem and the PostDetails page are no longer
the same, since the PostListItem is meant to be a simplified version of
the second one.

Right now the differences are that, in the details page,
the title of the post is not a link and that the buttons for editing and
removal are displayed.

The post list pages will only show some basic information on the post,
and will allow the user to update the score or enter the details page to
view all of its data and be able to update or delete it.