feat: Limit displayed characters in post list pages
Now the users cannot read the whole post content unless they go to the
details page. Instead, only a preview of 500 characters is shown.

If the post is longer than that, a "read more" link will appear to make
it clear that the view is missing part of the text.