feat: Scroll to comment form on edit
When the user clicked on the edit comment button, the comment form would
be fulfilled with its data. However, if the form was not in the user's
view, it was pretty easy to miss that something did actually happen.

I've changed things so, after populating the form with the comment
information, the page scrolls down automatically so the user see that it
has been updated with the selected comment data.