refactor: New utils method to sort by property
A new utils file has been created to contain useful methods that can be
reused all over the project. So far it contains only one method, which
receives an array of objects and the name of a property, and returns that same array
sorted by the value of that property. Additionally it can also receive a
third parameter with a direction, so we could potentially use it to
implement both ascending or descending order in our list pages.