refactor: Active comment in Redux store
The active comment was being stored in the CommentBox component state.
Since the aim of this project was to use Redux to keep all of the
application state (except for pieces of data that won't be reused anywhere
in the application, like the properties to hold the form values), the
active comment is now also present in the store.

The overall structure of the store is still messy, though.