fix: Create form reset after an edit
There was an issue where, if you went to the create form after having
updated another, the form was populated with the previously active post.

This issue has been addressed by checking if the component received a
postId as a prop and set the 'editMode' flag to false otherwise.

Also, if 'editMode' is set to true, then the form won't be filled until
the 'fetchPost' action has been completed, as to avoid loading data from a
wrong post if the store hasn't still being updated with the right post.