refactor: New subfolders in component
The 'components' folder is now slightly better organized, by including two
new subfolders. The first one, 'sidebar', contains a new 'SideBar'
component with the template for the sidebar of the application and all of
the components displayed in it (category list, sorting options and
application credits). The second, 'others', group more generic components
such as the loader and the error message view.