refactor: Post reducer simplified
The post reducer has been updated the same ways as the previous one, by
breaking it into smaller pieces and returning a final object with the
three available reducers inside.

An unused action has also been removed, and references to the post reducer
are updated to match the new store structure.