fix: Display right comment list in comment box
There was an issue with the comment list when entering page details for
the second time. The first time you enter a post you would see the right
set of comments. The second time you enter a different post, however, the
page would be displayed with the comments of the previous post you
watched.

The reason for that was that the action to fetch the comments would
trigger before the new post data was obtained, so the API was called with
the id of the previously active post.

Now the fetch comments action is being dispatched from the page details
page, right after the post data is obtained.

Also, since both activePost and the list of comments for it are stored in
the Redux store, both "CommentBox" and "CommentForm" are no longer
receiving any props via JSX. Instead, they read the values from the store.