fix: Replaced 'generate-uuid' package by 'uuid'
The 'generate-uuid' package seemed to be causing some trouble when doing
'npm install' after cloning the project from GitHub. Apparently it had
some kind of incompatibility with the rest of the packages installed,
because removing it solved the problem.

The PostsAPI file has been updated to use the 'uuid' package methods
instead of the 'generateUuid' function.